change float 32 to 64

+++ neo_finrl/joinquant/env_stock_jq.py
@@ -28,7 +28,7 @@ class StockTradingEnv():
         self.account = self.initial_account__reset
         self.day_npy = self.ary[self.day]
         self.prices = self.day_npy[[5*x+3 for x in range(self.stock_dim)]]
-        self.stocks = np.zeros(self.stock_dim, dtype=np.float32)  # multi-stack
+        self.stocks = np.zeros(self.stock_dim, dtype=np.float64)  # multi-stack
 
         self.total_asset = self.account + (self.prices[:self.stock_dim] * self.stocks).sum()
         self.episode_return = 0.0  # Compatibility for ElegantRL 2020-12-21
@@ -44,7 +44,7 @@ class StockTradingEnv():
 
     def reset(self):
         self.account = self.initial_account
-        self.stocks = np.zeros(self.stock_dim, dtype=np.float32)
+        self.stocks = np.zeros(self.stock_dim, dtype=np.float64)
         self.prices = self.day_npy[[5*x+3 for x in range(self.stock_dim)]]
         self.total_asset = self.account + (self.prices[:self.stock_dim] * self.stocks).sum()
         # total_asset = account + (adjcp * stocks).sum()
@@ -55,7 +55,7 @@ class StockTradingEnv():
 
         state = np.hstack((self.account * 2 ** -16,
                            self.day_npy * 2 ** -8,
-                           self.stocks * 2 ** -12,), ).astype(np.float32)
+                           self.stocks * 2 ** -12,), ).astype(np.float64)
         return state
 
     def step(self, action):
@@ -83,7 +83,7 @@ class StockTradingEnv():
 
         state = np.hstack((self.account * 2 ** -16,
                            self.day_npy * 2 ** -8,
-                           self.stocks * 2 ** -12,), ).astype(np.float32)
+                           self.stocks * 2 ** -12,), ).astype(np.float64)
 
         next_total_asset = self.account + (self.prices[:self.stock_dim] * self.stocks).sum()
         reward = (next_total_asset - self.total_asset) * 2 ** -16  # notice scaling!
